--- src/Nethermind/Nethermind.Blockchain/BlockHashInStateExtension.cs
@@ -1,16 +1,0 @@
-// SPDX-FileCopyrightText: 2024 Demerzel Solutions Limited
-// SPDX-License-Identifier: LGPL-3.0-only
-
-using System;
-using System.IO;
-using Nethermind.Blockchain.Find;
-using Nethermind.Core;
-using Nethermind.Core.Specs;
-using Nethermind.State;
-
-namespace Nethermind.Blockchain;
-
-public static class BlockHashInStateExtension
-{
-
-}
